--- apps/prairielearn/src/pages/studentAssessmentHomework/studentAssessmentHomework.ejs
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <%- include('../partials/head'); %>
-  </head>
-  <body>
-    <%- include('../partials/navbar', {navPage: ''}); %>
-    <main id="content" class="container">
-      <div class="card mb-4">
-        <div class="card-header bg-primary text-white">
-          <%= assessment_set.abbreviation %><%= assessment.number %>: <%= assessment.title %>
-          <i class="fas fa-users"></i>
-        </div>
-        <div class="card-body">
-          <p class="lead text-center">
-            <strong><%= assessment_set.name %> <%= assessment.number %>: <%= assessment.title %></strong> for <strong><%= course.short_name %></strong>
-          </p>
-          <%- include('../partials/studentGroupControls.ejs'); %>
-        </div>
-      </div>
-    </main>
-  </body>
-</html>
